pyproject.toml
[tool.poetry]
name = "exp-app"
version = "0.1.0"
description = "Expenses app with frontend(React) and backend(FastApi)"
authors = ["DmitriiV84 <dmitriiv84@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beanie = "^1.25.0"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"
pydantic-extra-types = "^2.6.0"
gunicorn = "^21.2.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
odmantic = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
mypy = "^1.9.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
